RavenDB-21763 fix (move server settings to react): Code review fixes
